backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication
djangorestframework-simplejwt==5.3.1
django-allauth==0.60.1
dj-rest-auth==5.0.2

# Media & Storage
Pillow==10.2.0
django-storages==1.14.2
boto3==1.34.14

# Caching & Async
redis==5.0.1
django-redis==5.4.0
celery==5.3.6

# API Documentation
drf-spectacular==0.27.0

# Utilities
python-decouple==3.8
django-extensions==3.2.3
django-debug-toolbar==4.2.0

# Email
django-ses==3.5.2

# Search
django-elasticsearch-dsl==8.0
elasticsearch==8.11.1

# Admin Enhancement
django-jazzmin==2.6.0

# Security
django-ratelimit==4.1.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0

# Code Quality
black==23.12.1
flake8==7.0.0
isort==5.13.2

# Production
gunicorn==21.2.0
whitenoise==6.6.0
